Added low bit counting for auto-recalibration.
This increases the minimum bit period for receiving from 33 to 34
clocks, but that is for precise mid-bit interval sampling. I expect 33
clocks will still work in practice.

The current approach does not work if an interbyte timeout occurs (which
is expected to be relatively rare). I see how to fix this, but I may
prioritize detecting the byte period.