Major changes to receive code.
Discovered a bug in the receiving code -- rxPrevByte was not being reset
on interbyte timeout, which would corrupt lowBitCount. Seeing no way to
fit a reset in, I had to rewrite the receive code so that it calculates
lowBitCount on the current byte instead. Made several other changes,
such as differentiating between framing and parsing errors. Started on
register optimization.